new cache module with chunk and packrat caches
The low-level chunk cache maps input position ranges to "chunk" arrays
which are allocated as needed. This is faster than a hash-table and
smaller than an array equal to the size of the input in terms of
up-front memory usage.

The "head cache" only uses input positions as keys and can thus be
implemented directly using a chunk cache.

The packrat cache uses pairs of input positions and rule names as
keys and thus needs a second level on top of the chunk cache. This
second level is implemented as an adaptive lookup data structure that
uses either a single cons, a short alist or a hash-table, depending on
the number of rules.