automatically compile simple rules without caching
For a rule with a simple expression, if the semantics do not depend on
caching, compile the rule such that it does not perform a packrat
cache lookup and update.

This is an optimization since, for a simple expression, a cache
lookup (let alone lookup and update) is more expensive than simply
evaluating the expression.